backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
dnspython==2.8.0
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.128.0
google-auth==2.47.0
greenlet==3.3.0
h11==0.16.0
idna==3.11
passlib==1.7.4
psycopg==3.3.2
psycopg-binary==3.3.2
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
pyjwt==2.10.1
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
soupsieve==2.8.1
sqlalchemy==2.0.45
sqlmodel==0.0.31
starlette==0.50.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
